student support/requirements.txt
flask
mysql-connector-python
pandas
scikit-learn
nltk
joblib
deep-translator

